# src/openapi.yaml
openapi: 3.1.0
info:
  title: Branch Bank API
  version: 1.0.0
  description: |
    The branch bank API provides user-facing operations for account lifecycle management
    and fund transfers in the distributed banking system.

    This API allows users to register with a branch bank, create accounts with chosen
    currencies, perform unauthenticated account lookups to verify ownership, and initiate
    same-bank and cross-bank fund transfers.

    ## Account Number Semantics

    ### Account Number Format
    - Account numbers are exactly 8 characters long
    - The first 3 characters represent the bank identifier (matching the bankId from the central bank registry)
    - The last 5 characters are unique within the bank and generated by the branch bank
    - Format: `{BANK_PREFIX}{UNIQUE_SUFFIX}` where BANK_PREFIX is 3 uppercase letters/digits and UNIQUE_SUFFIX is 5 alphanumeric characters

    Example: "EST12345" where:
    - "EST" = bank prefix (e.g., Estonia Commercial Bank)
    - "12345" = unique account identifier within that bank

    ### Account Lookup Behavior
    - GET /accounts/{accountNumber} is an unauthenticated endpoint
    - Returns the owner's full name when the account exists
    - Returns 404 Not Found when the account does not exist
    - No authentication is required for lookup, making it possible to verify account ownership before transfers

    This contract is implementation-agnostic and serves as the source of truth for interoperability
...

  contact:
...
servers:
...
paths:
  /users:
    post:
      tags:
        - Users
      summary: Register a new user
      description: |
        Registers a new user with the branch bank. Users must be registered before they can create accounts.

        **Registration Flow:**
        - User provides their full name and optional email address
        - Branch bank assigns a unique userId
        - User can immediately create accounts after registration

        **User ID Format:**
        User IDs are unique within the branch bank and follow the pattern: `user-{UUID}` where UUID is a standard UUID v4 format.

        **Note:** This endpoint does not require authentication for initial registration, as it is the first step in creating a banking relationship.
      operationId: registerUser
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegistrationRequest'
            examples:
              userRegistration:
                summary: Standard user registration
                value:
                  fullName: "Jane Doe"
                  email: "jane.doe@example.com"
              minimalRegistration:
                summary: Minimal user registration (email optional)
                value:
                  fullName: "John Smith"
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserRegistrationResponse'
              examples:
                successfulRegistration:
                  summary: User registered
                  value:
                    userId: "user-550e8400-e29b-41d4-a716-446655440000"
                    fullName: "Jane Doe"
                    email: "jane.doe@example.com"
                    createdAt: "2026-04-02T11:15:00Z"
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                invalidRequest:
                  summary: Missing required fields
                  value:
                    code: "INVALID_REQUEST"
                    message: "Full name is required"
        '409':
          description: User already registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                duplicateUser:
                  summary: User with same email already exists
                  value:
                    code: "DUPLICATE_USER"
                    message: "A user with this email address is already registered"
  /users/{userId}:
    get:
      tags:
        - Users
      summary: Get user profile
      description: Returns user information including their accounts.
      operationId: getUser
      security: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            example: "user-550e8400-e29b-41d4-a716-446655440000"
      responses:
        '200':
          description: User found
          content:
            application/json:
              schema:
                type: object
                properties:
                  userId:
                    type: string
                  fullName:
                    type: string
                  email:
                    type: string
                  createdAt:
                    type: string
                    format: date-time
                  accounts:
                    type: array
                    items:
                      type: object
                      properties:
                        account_number:
                          type: string
                        currency:
                          type: string
                        balance:
                          type: string
                        created_at:
                          type: string
        '404':
          description: User not found
  /users/{userId}/accounts:
    get:
      tags:
        - Accounts
      summary: List user accounts with balances
      description: Returns all accounts owned by the user with current balances. Requires authentication.
      operationId: listUserAccounts
      security:
        - BearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Accounts listed
          content:
            application/json:
              schema:
                type: object
                properties:
                  accounts:
                    type: array
                    items:
                      type: object
                      properties:
                        account_number:
                          type: string
                        currency:
                          type: string
                        balance:
                          type: string
                        created_at:
                          type: string
        '401':
          description: Unauthorized
    post:
      tags:
        - Accounts
      summary: Create a new account for a user
      description: |
        Creates a new bank account for an existing user with the specified currency.

        **Account Creation Flow:**
        - User must be registered (valid userId required)
        - User specifies the desired currency for the account
        - Branch bank generates a unique 8-character account number
        - Account number format: `{BANK_PREFIX}{UNIQUE_SUFFIX}` (3 + 5 characters)
        - Initial balance is set to 0.00

        **Account Number Rules:**
        - Account numbers must be exactly 8 characters
        - The first 3 characters must match the bank's prefix (assigned by the central bank)
        - The last 5 characters are unique within the bank and auto-generated
        - Account numbers are case-insensitive in storage but should be uppercase in responses

        **Currency Support:**
        Branch banks typically support major currencies. The requested currency must be one that the bank supports.

        **Note:** This endpoint requires the user to be authenticated (e.g., via session token or API key).
      operationId: createAccount
      security:
        - BearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          description: The unique identifier of the user
          schema:
            type: string
            pattern: '^user-[0-9a-f]{8}-[0-9a-f]{4}-4[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$'
            example: "user-550e8400-e29b-41d4-a716-446655440000"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountCreationRequest'
            examples:
              eurosAccount:
                summary: Create Euro account
                value:
                  currency: "EUR"
              dollarsAccount:
                summary: Create US Dollar account
                value:
                  currency: "USD"
      responses:
        '201':
          description: Account created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountCreationResponse'
              examples:
                successfulCreation:
                  summary: Euro account created
                  value:
                    accountNumber: "EST12345"
                    ownerId: "user-550e8400-e29b-41d4-a716-446655440000"
                    currency: "EUR"
                    balance: "0.00"
                    createdAt: "2026-04-02T11:15:15Z"
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                invalidRequest:
                  summary: Missing or invalid currency
                  value:
                    code: "INVALID_REQUEST"
                    message: "Currency is required and must be a valid ISO 4217 code"
                invalidCurrency:
                  summary: Unsupported currency
                  value:
                    code: "UNSUPPORTED_CURRENCY"
                    message: "Currency 'XXX' is not supported by this bank"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                unauthorized:
                  summary: Authentication required
                  value:
                    code: "UNAUTHORIZED"
                    message: "Authentication is required to create an account"
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                userNotFound:
                  summary: User does not exist
                  value:
                    code: "USER_NOT_FOUND"
                    message: "User with ID 'user-550e8400-e29b-41d4-a716-446655440000' not found"
  /accounts/{accountNumber}:
    get:
      tags:
        - Accounts
      summary: Look up account by account number
      description: |
        Retrieves account ownership information for a specific account number.

        **Unauthenticated Access:**
        - This endpoint does NOT require authentication
        - Anyone can look up an account number to verify ownership
        - This enables pre-transfer verification and account validation

        **Response Behavior:**
        - Returns the owner's full name when the account exists
        - Returns 404 Not Found when the account does not exist

        **Use Cases:**
        - Verify account ownership before initiating transfers
        - Validate account numbers provided by users
        - Debug and troubleshooting

        **Privacy Note:**
        Only the owner's name is returned, not their email, balance, or transaction history. This balances
        utility (allowing account verification) with privacy (protecting sensitive financial information).

        **Account Number Format:**
        Account numbers must be exactly 8 characters and match the pattern: `[A-Z0-9]{3}[A-Z0-9]{5}`.
        The first 3 characters identify the bank, and the last 5 are the unique account identifier.
      operationId: lookupAccount
      security: []
      parameters:
        - name: accountNumber
          in: path
          required: true
          description: |
            The 8-character account number to look up.

            Format: `{BANK_PREFIX}{UNIQUE_SUFFIX}`
            - BANK_PREFIX: 3 characters (the bank's identifier)
            - UNIQUE_SUFFIX: 5 characters (unique within the bank)

            Example: "EST12345" where "EST" is the bank prefix
          schema:
            type: string
            pattern: '^[A-Z0-9]{8}$'
            example: "EST12345"
      responses:
        '200':
          description: Account found - owner information returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountLookupResponse'
              examples:
                accountFound:
                  summary: Account exists
                  value:
                    accountNumber: "EST12345"
                    ownerName: "Jane Doe"
                    currency: "EUR"
        '404':
          description: Account not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                accountNotFound:
                  summary: Account does not exist
                  value:
                    code: "ACCOUNT_NOT_FOUND"
                    message: "Account with number 'EST12345' not found"
        '400':
          description: Invalid account number format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                invalidAccountNumber:
                  summary: Malformed account number
                  value:
                    code: "INVALID_ACCOUNT_NUMBER"
                    message: "Account number must be exactly 8 characters"
  /users/{userId}/transfers:
    get:
      tags:
        - Transfers
      summary: Get transfer history for user
      description: Returns all transfers where the user's accounts are source or destination.
      operationId: getUserTransfers
      security:
        - BearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Transfer history
          content:
            application/json:
              schema:
                type: object
                properties:
                  transfers:
                    type: array
                    items:
                      $ref: '#/components/schemas/TransferResponse'
        '401':
          description: Unauthorized
  /transfers:
    post:
      tags:
        - Transfers
      summary: Initiate a fund transfer
      description: |
        Initiates a fund transfer from a source account to a destination account.

        **Routing (D-01):**
        Branch bank determines routing by inspecting destination account's bank prefix (first 3 characters).
        Same-bank transfers execute directly; cross-bank transfers route via POST /transfers/receive.

        **Idempotency (D-05):**
        Each transfer requires a unique transferId (UUID format). Duplicate transferId values are rejected.

        **Currency Conversion (XFER-04):**
        Cross-bank transfers convert currency at current exchange rate from central bank (D-03).

        **Cross-Bank Authentication (D-04):**
        Inter-bank transfers use JWT signed with source bank's private key (ES256 algorithm).

        **Resilient Delivery (Phase 3):**

        **Pending Transfers (XFER-06):**
        When destination bank is unavailable during cross-bank transfer:
        - Status is set to "pending" with immediate fund deduction from source account
        - Transfer is queued with exponential backoff retry (1m → 2m → 4m → ... → 1h)
        - Funds remain locked until transfer succeeds, fails permanently, or times out
        - Duplicate transferId submission returns 409: TRANSFER_ALREADY_PENDING

        **Central Bank Unavailability (XFER-05):**
        When central bank directory service is unavailable:
        - Branch banks use cached bank directory data for routing
        - Returns 503: CENTRAL_BANK_UNAVAILABLE (transfer proceeds using cache)

        **Timeout Handling (XFER-07):**
        Pending transfers expire after 4 hours:
        - Status changes to "failed_timeout"
        - Funds are automatically refunded to source account
        - Modification attempts return 423: TRANSFER_TIMEOUT
      operationId: initiateTransfer
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransferRequest'
            examples:
              sameBankTransfer:
                summary: Same-bank transfer
                value:
                  transferId: "550e8400-e29b-41d4-a716-446655440000"
                  sourceAccount: "EST12345"
                  destinationAccount: "EST54321"
                  amount: "100.00"
              crossBankTransfer:
                summary: Cross-bank transfer
                value:
                  transferId: "660e9511-f30c-52e5-b827-557766551111"
                  sourceAccount: "EST12345"
                  destinationAccount: "LAT54321"
                  amount: "100.00"
      responses:
        '201':
          description: Transfer initiated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransferResponse'
              examples:
                sameBankCompleted:
                  summary: Same-bank transfer completed
                  value:
                    transferId: "550e8400-e29b-41d4-a716-446655440000"
                    status: "completed"
                    sourceAccount: "EST12345"
                    destinationAccount: "EST54321"
                    amount: "100.00"
                    timestamp: "2026-04-08T12:00:05Z"
                crossBankCompleted:
                  summary: Cross-bank transfer completed
                  value:
                    transferId: "660e9511-f30c-52e5-b827-557766551111"
                    status: "completed"
                    sourceAccount: "EST12345"
                    destinationAccount: "LAT54321"
                    amount: "100.00"
                    convertedAmount: "85.00"
                    exchangeRate: "0.850000"
                    rateCapturedAt: "2026-04-08T12:00:00Z"
                    timestamp: "2026-04-08T12:00:10Z"
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Account not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Duplicate transfer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                duplicateTransfer:
                  summary: Transfer ID already used
                  value:
                    code: "DUPLICATE_TRANSFER"
                    message: "A transfer with ID '550e8400-e29b-41d4-a716-446655440000' already exists"
                transferAlreadyPending:
                  summary: Transfer already pending
                  value:
                    code: "TRANSFER_ALREADY_PENDING"
                    message: "Transfer with ID '550e8400-e29b-41d4-a716-446655440000' is already pending. Cannot submit duplicate transfer."
        '422':
          description: Unprocessable entity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                insufficientFunds:
                  summary: Insufficient funds
                  value:
                    code: "INSUFFICIENT_FUNDS"
                    message: "Insufficient funds in source account"
        '503':
          description: Service unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                destinationBankUnavailable:
                  summary: Destination bank unavailable
                  value:
                    code: "DESTINATION_BANK_UNAVAILABLE"
                    message: "Destination bank is temporarily unavailable. Transfer has been queued for retry."
                centralBankUnavailable:
                  summary: Central bank unavailable
                  value:
                    code: "CENTRAL_BANK_UNAVAILABLE"
                    message: "Central bank is temporarily unavailable. Using cached directory data for routing."
  /transfers/receive:
    post:
      tags:
        - Transfers
      summary: Receive cross-bank transfer
      description: |
        Receives and processes a cross-bank transfer initiated by another bank.
        Follows locked design decisions D-02 (inter-bank endpoint), D-04 (JWT authentication),
        and DEC-05 (HTTP body contains only JWT field).
      operationId: receiveInterBankTransfer
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InterBankTransferRequest'
            examples:
              interBankTransfer:
                summary: Cross-bank transfer JWT
                value:
                  jwt: "eyJhbGciOiJFUzI1NiIsInR5cCI6IkpXVCJ9.eyJ0cmFuc2ZlcklkIjoiNTUwZTg0MDAtZTI5Yi00MWQ0LWE3MTYtNDQ2NjU1NDQwMDAwIiwic291cmNlQWNjb3VudCI6IkVTVDEyMzQ1IiwiZGVzdGluYXRpb25BY2NvdW50IjoiTEFUNTU0MzIxIiwiYW1vdW50IjoiMTAwLjAwIiwic291cmNlQmFua0lkIjoiRVNUMDEiLCJkZXN0aW5hdGlvbkJhbmtJZCI6IkxBVDAwMiIsInRpbWVzdGFtcCI6IjIwMjYtMDQtMDhUMTI6MDA6MDBaIiwibm9uY2UiOiJhMWIyYzNkNGU1ZjZnN2g4In0.signature"
      responses:
        '200':
          description: Transfer received successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InterBankTransferResponse'
              examples:
                transferAccepted:
                  summary: Transfer accepted
                  value:
                    transferId: "550e8400-e29b-41d4-a716-446655440000"
                    status: "completed"
                    destinationAccount: "LAT54321"
                    amount: "100.00"
                    timestamp: "2026-04-08T12:00:15Z"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Account not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /transfers/{transferId}:
    get:
      tags:
        - Transfers
      summary: Get transfer status
      description: |
        Retrieves the current status and details of a transfer.
        Decision DEC-04: Include this endpoint in Phase 2.
      operationId: getTransferStatus
      security:
        - BearerAuth: []
      parameters:
        - name: transferId
          in: path
          required: true
          description: The unique identifier of the transfer
          schema:
            type: string
            format: uuid
            example: "550e8400-e29b-41d4-a716-446655440000"
      responses:
        '200':
          description: Transfer status retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransferStatusResponse'
              examples:
                pendingTransfer:
                  summary: Transfer pending (destination bank unavailable)
                  value:
                    transferId: "770e4033-j50e-64i7-f370-001122334455"
                    status: "pending"
                    sourceAccount: "EST12345"
                    destinationAccount: "LAT22222"
                    amount: "500.00"
                    pendingSince: "2026-04-08T14:00:00Z"
                    nextRetryAt: "2026-04-08T14:01:00Z"
                    retryCount: 1
                    timestamp: "2026-04-08T14:00:05Z"
                completedTransfer:
                  summary: Transfer completed
                  value:
                    transferId: "550e8400-e29b-41d4-a716-446655440000"
                    status: "completed"
                    sourceAccount: "EST12345"
                    destinationAccount: "LAT54321"
                    amount: "100.00"
                    timestamp: "2026-04-08T12:00:15Z"
                failedTransfer:
                  summary: Transfer failed
                  value:
                    transferId: "880e3944-j64f-85h8-e261-001122334455"
                    status: "failed"
                    sourceAccount: "EST11111"
                    destinationAccount: "LAT22222"
                    amount: "1000.00"
                    timestamp: "2026-04-08T12:01:00Z"
                    errorMessage: "Insufficient funds in source account"
                timeoutFailedTransfer:
                  summary: Transfer failed due to timeout
                  value:
                    transferId: "880e3944-j64f-85h8-e261-001122334455"
                    status: "failed_timeout"
                    sourceAccount: "EST11111"
                    destinationAccount: "LAT22222"
                    amount: "1000.00"
                    timestamp: "2026-04-08T18:00:00Z"
                    errorMessage: "Transfer timed out after 4 hours. Funds refunded to source account."
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Transfer not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                transferNotFound:
                  summary: Transfer not found
                  value:
                    code: "TRANSFER_NOT_FOUND"
                    message: "Transfer with ID '550e8400-e29b-41d4-a716-446655440000' not found"
        '423':
          description: Transfer is locked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                transferTimeout:
                  summary: Transfer timed out
                  value:
                    code: "TRANSFER_TIMEOUT"
                    message: "Transfer has timed out and cannot be modified or retried. Status is failed_timeout with refund processed."
components:
  schemas:
    UserRegistrationRequest:
      type: object
      required:
        - fullName
      properties:
        fullName:
          type: string
          description: The full name of the user (used in account lookups)
          minLength: 2
          maxLength: 200
          example: "Jane Doe"
        email:
          type: string
          format: email
          description: Optional email address for the user (used for notifications)
          maxLength: 255
          example: "jane.doe@example.com"
    UserRegistrationResponse:
      type: object
      required:
        - userId
        - fullName
        - createdAt
      properties:
        userId:
          type: string
          description: The unique identifier assigned to this user
          pattern: '^user-[0-9a-f]{8}-[0-9a-f]{4}-4[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$'
          example: "user-550e8400-e29b-41d4-a716-446655440000"
        fullName:
          type: string
          description: The full name of the registered user
          example: "Jane Doe"
        email:
          type: string
          format: email
          description: The email address of the user (if provided)
          example: "jane.doe@example.com"
        createdAt:
          type: string
          format: date-time
          description: |
            The timestamp when the user was registered.

            **Note:** This timestamp is stored on the branch bank server and reflects
            the server's clock at the time of registration.
          example: "2026-04-02T11:15:00Z"
    AccountCreationRequest:
      type: object
      required:
        - currency
      properties:
        currency:
          type: string
          description: |
            The ISO 4217 currency code for the account.

            **Supported Currencies:**
            Branch banks typically support major currencies such as EUR, USD, GBP, etc.

            **Validation:**
            The currency must be a valid 3-letter ISO 4217 code and supported by the specific branch bank.

            **Examples:**
            - EUR: Euro
            - USD: United States Dollar
            - GBP: British Pound
          pattern: '^[A-Z]{3}$'
          example: "EUR"
...
    AccountLookupResponse:
      type: object
      required:
        - accountNumber
        - ownerName
        - currency
      properties:
        accountNumber:
          type: string
          description: The 8-character account number that was looked up
          pattern: '^[A-Z0-9]{8}$'
          example: "EST12345"
        ownerName:
          type: string
          description: |
            The full name of the account owner.

            **Privacy Note:**
            Only the owner's name is returned. Email addresses, balances, transaction history,
            and other sensitive information are NOT exposed through this unauthenticated endpoint.

            This design allows account verification transfers while protecting user privacy.
          example: "Jane Doe"
        currency:
          type: string
          description: The ISO 4217 currency code for this account
          pattern: '^[A-Z]{3}$'
          example: "EUR"
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: A machine-readable error code
          example: "ACCOUNT_NOT_FOUND"
        message:
          type: string
          description: A human-readable error message
          example: "Account with number 'EST12345' not found"
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT bearer token for authenticated operations.

        **Required For:**
        - Account creation (POST /users/{userId}/accounts)
        - Transfer initiation (POST /transfers)
        - Transfer status lookup (GET /transfers/{transferId})

        **Not Required For:**
        - User registration (POST /users) - initial signup is open
        - Account lookup (GET /accounts/{accountNumber}) - unauthenticated for verification
        - Inter-bank transfer reception (POST /transfers/receive) - authenticated via bank-signed JWT (D-04)

        **Token Format:**
        Tokens are signed JWTs containing the userId and expiration time.